+++ components/feedback/test/index.test.js
@@ -1,4 +1,4 @@
-import { expect, fixture, html } from '@open-wc/testing';
+import { expect, fixture } from '@open-wc/testing';
 /* this test file can be run on command line with npm run test
    or with visual debug via npm run test:visual */
 
@@ -7,10 +7,13 @@ import '../dist/index.js';
 describe('CAGOV Page feedback', function unitTest() {
   this.timeout(5000);
   it('works', async () => {
-    const el = await fixture(html`<cagov-feedback
-      data-endpoint-url="my-endpoint-url"
-    >
-    </cagov-feedback>`);
+    const response = await fetch('../template.html');
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
+    }
+
+    const startHTML = await response.text();
+    const el = await fixture(startHTML);
 
     // verify textarea is not display block
     expect(el.querySelector('.feedback-form-add').style.display).to.equal('');